@@ -8,144 +8,137 @@
     "Anna Bodnia <anna.bodnia@gmail.com>",
     "Buu Nguyen <buunguyen@gmail.com>",
     "Josh Ponelat <jponelat@gmail.com>",
-    "Kyle Shockey <kyleshockey1@gmail.com>",
+    "Kyle Shockey <kyleshockey@gmail.com>",
     "Robert Barnwell <robert@robertismy.name>",
     "Sahar Jafari <shr.jafari@gmail.com>"
   ],
   "license": "Apache-2.0",
   "scripts": {
-    "build": "npm run build-core && npm run build-bundle && npm run build-standalone",
+    "build": "run-p --aggregate-output build-core build-bundle build-standalone",
     "build-bundle": "webpack --config webpack-dist-bundle.config.js --colors",
     "build-core": "webpack --config webpack-dist.config.js --colors",
     "build-standalone": "webpack --config webpack-dist-standalone.config.js --colors",
     "predev": "npm install",
-    "dev": "npm-run-all --parallel hot-server open-localhost",
+    "dev": "npm-run-all --parallel hot-server",
     "watch": "webpack --config webpack-watch.config.js --watch --progress",
-    "open-localhost": "node -e 'require(\"open\")(\"http://localhost:3200\")'",
     "hot-server": "webpack-dev-server --host 0.0.0.0 --config webpack-hot-dev-server.config.js --inline --hot --progress --content-base dev-helpers/",
     "deps-license": "license-checker --production --csv --out $npm_package_config_deps_check_dir/licenses.csv && license-checker --development --csv --out $npm_package_config_deps_check_dir/licenses-dev.csv",
     "deps-size": "webpack -p --config webpack.check.js --json | webpack-bundle-size-analyzer >| $npm_package_config_deps_check_dir/sizes.txt",
-    "deps-check": "npm run deps-license && npm run deps-size",
+    "deps-check": "run-s deps-license deps-size",
     "lint": "eslint --cache --ext '.js,.jsx' src test",
     "lint-errors": "eslint --cache --quiet --ext '.js,.jsx' src test",
     "lint-fix": "eslint --cache --ext '.js,.jsx' src test --fix",
-    "test": "npm run lint-errors && npm run just-test-in-node",
-    "test-in-node": "npm run lint-errors && npm run just-test-in-node",
-    "just-test": "karma start --config karma.conf.js",
-    "just-test-in-node": "mocha --recursive --compilers js:babel-core/register test/core test/components test/bugs test/swagger-ui-dist-package test/xss",
+    "test": "run-s just-test-in-node e2e-cypress lint-errors",
+    "test-in-node": "run-s lint-errors just-test-in-node",
+    "just-test-in-node": "mocha --require test/setup.js --recursive --compilers js:babel-core/register --require source-map-support test/core test/components test/bugs test/swagger-ui-dist-package test/xss",
     "just-check-coverage": "nyc npm run just-test-in-node",
-    "test-e2e": "sleep 3 && nightwatch test/e2e/scenarios/ --config test/e2e/nightwatch.json",
-    "e2e-initial-render": "nightwatch test/e2e/scenarios/ --config test/e2e/nightwatch.json --group initial-render",
-    "mock-api": "json-server --watch test/e2e/db.json --port 3204",
-    "e2e": "npm-run-all --parallel -r hot-server mock-api test-e2e"
+    "test-e2e-cypress": "cypress run",
+    "test-e2e-selenium": "sleep 3 && nightwatch test/e2e-selenium/scenarios/ --config test/e2e-selenium/nightwatch.json",
+    "e2e-initial-render": "nightwatch test/e2e-selenium/scenarios/ --config test/e2e-selenium/nightwatch.json --group initial-render",
+    "mock-api": "json-server --watch test/e2e-selenium/db.json --port 3204",
+    "hot-e2e-cypress-server": "webpack-dev-server --port 3230 --content-base test/e2e-cypress/static --host 0.0.0.0 --config webpack-hot-dev-server.config.js --inline --hot --progress",
+    "hot-e2e-selenium-server": "webpack-dev-server --port 3230 --content-base test/e2e-selenium/helpers --host 0.0.0.0 --config webpack-hot-dev-server.config.js --inline --hot --progress",
+    "e2e-cypress": "run-p -r hot-e2e-cypress-server mock-api test-e2e-cypress",
+    "e2e-selenium": "run-p -r hot-e2e-selenium-server mock-api test-e2e-selenium",
+    "open-static": "node -e 'require(\"open\")(\"http://localhost:3002\")'",
+    "serve-static": "http-server dist/ -i -a 0.0.0.0 -p 3002",
+    "start": "npm-run-all --parallel serve-static open-static"
   },
   "dependencies": {
     "@braintree/sanitize-url": "^2.0.2",
+    "@kyleshockey/js-yaml": "^1.0.1",
     "base64-js": "^1.2.0",
-    "brace": "0.7.0",
     "classnames": "^2.2.5",
-    "commonmark": "^0.28.1",
     "core-js": "^2.5.1",
     "css.escape": "1.5.1",
-    "deep-extend": "0.4.1",
-    "expect": "1.20.2",
-    "getbase": "^2.8.2",
+    "deep-extend": "0.5.1",
+    "dompurify": "^1.0.4",
     "ieee754": "^1.1.8",
     "immutable": "^3.x.x",
-    "js-yaml": "^3.5.5",
-    "lodash": "4.17.2",
-    "matcher": "^0.1.2",
-    "memoizee": "0.4.1",
-    "promise-worker": "^1.1.1",
+    "js-file-download": "^0.4.1",
+    "lodash": "^4.17.10",
+    "memoizee": "^0.4.12",
     "prop-types": "^15.5.10",
     "react": "^15.6.2",
     "react-addons-perf": "^15.4.0",
-    "react-addons-shallow-compare": "0.14.8",
-    "react-addons-test-utils": "^15.6.2",
-    "react-collapse": "^4.0.3",
+    "react-debounce-input": "^3.2.0",
     "react-dom": "^15.6.2",
-    "react-height": "^2.0.0",
-    "react-hot-loader": "1.3.1",
     "react-immutable-proptypes": "2.1.0",
     "react-immutable-pure-component": "^1.1.1",
-    "react-markdown": "^2.5.0",
+    "react-inspector": "^2.3.0",
     "react-motion": "^0.5.2",
-    "react-object-inspector": "0.2.1",
     "react-redux": "^4.x.x",
-    "react-split-pane": "0.1.70",
     "redux": "^3.x.x",
     "redux-immutable": "3.0.8",
-    "redux-logger": "*",
     "remarkable": "^1.7.1",
-    "reselect": "2.5.3",
-    "sanitize-html": "^1.14.1",
-    "scroll-to-element": "^2.0.0",
-    "serialize-error": "2.0.0",
-    "shallowequal": "0.2.2",
-    "swagger-client": "^3.4.6",
+    "reselect": "^2.5.4",
+    "serialize-error": "^2.1.0",
+    "swagger-client": "^3.8.19",
     "url-parse": "^1.1.8",
-    "whatwg-fetch": "0.11.1",
-    "worker-loader": "^0.7.1",
     "xml": "1.0.1",
     "xml-but-prettier": "^1.0.1",
-    "yaml-js": "0.2.0",
-    "zenscroll": "4.0.0"
+    "zenscroll": "^4.0.2"
   },
   "devDependencies": {
-    "autoprefixer": "7.1.1",
+    "autoprefixer": "^8.4.1",
     "babel-core": "^6.23.1",
     "babel-eslint": "^7.1.1",
     "babel-loader": "^7.1.0",
     "babel-plugin-module-alias": "^1.6.0",
+    "babel-plugin-transform-es2015-constants": "^6.1.4",
+    "babel-plugin-transform-react-remove-prop-types": "^0.4.13",
     "babel-plugin-transform-runtime": "^6.23.0",
     "babel-preset-es2015": "^6.22.0",
     "babel-preset-es2015-ie": "^6.6.2",
     "babel-preset-react": "^6.23.0",
     "babel-preset-stage-0": "^6.22.0",
     "babel-runtime": "^6.23.0",
-    "chromedriver": "^2.30.1",
+    "bundlesize": "^0.17.0",
+    "chromedriver": "^2.38.3",
     "copy-webpack-plugin": "^4.0.1",
-    "css-loader": "0.28.4",
-    "deep-extend": "^0.5.0",
-    "deepmerge": "^1.3.2",
+    "css-loader": "^0.28.11",
+    "cypress": "^3.1.0",
+    "dedent": "^0.7.0",
+    "deepmerge": "^2.1.0",
     "enzyme": "^2.7.1",
     "eslint": "^4.1.1",
-    "eslint-plugin-import": "^2.6.0",
+    "eslint-plugin-import": "^2.13.0",
     "eslint-plugin-mocha": "^4.11.0",
-    "eslint-plugin-react": "^7.1.0",
-    "extract-text-webpack-plugin": "^2.1.2",
-    "file-loader": "0.11.2",
+    "eslint-plugin-react": "^7.10.0",
+    "expect": "^1.20.2",
+    "extract-text-webpack-plugin": "^3.0.2",
+    "file-loader": "^1.1.11",
     "git-describe": "^4.0.1",
-    "imports-loader": "0.7.1",
-    "json-loader": "0.5.4",
-    "json-server": "^0.11.0",
-    "karma": "^1.7.0",
-    "karma-chrome-launcher": "^2.2.0",
-    "karma-mocha": "^1.3.0",
-    "karma-sourcemap-loader": "^0.3.7",
-    "karma-webpack": "2.0.3",
-    "less": "2.7.2",
-    "license-checker": "^11.0.0",
-    "mocha": "^3.4.2",
+    "http-server": "^0.11.1",
+    "imports-loader": "^0.8.0",
+    "jsdom": "^11.10.0",
+    "json-loader": "^0.5.7",
+    "json-server": "^0.12.2",
+    "less": "^3.0.2",
+    "license-checker": "^19.0.0",
+    "mocha": "^5.1.1",
     "nightwatch": "^0.9.16",
     "node-sass": "^4.5.0",
-    "npm-run-all": "4.0.2",
+    "npm-run-all": "^4.1.2",
     "null-loader": "0.1.1",
     "nyc": "^11.3.0",
     "open": "0.0.5",
-    "postcss-loader": "2.0.6",
+    "postcss-loader": "^2.1.5",
     "raw-loader": "0.5.1",
-    "react-hot-loader": "^1.3.1",
     "react-test-renderer": "^15.5.4",
     "rimraf": "^2.6.0",
-    "sass-loader": "^6.0.2",
-    "selenium-server-standalone-jar": "3.4.0",
-    "standard": "^10.0.2",
+    "sass-loader": "^7.0.1",
+    "selenium-server-standalone-jar": "3.12.0",
+    "standard": "^11.0.1",
     "standard-loader": "^6.0.1",
-    "style-loader": "0.18.2",
-    "url-loader": "0.5.9",
-    "webpack": "^2.6.1",
-    "webpack-bundle-size-analyzer": "^2.5.0"
+    "style-loader": "^0.21.0",
+    "tachyons-sass": "^4.9.2",
+    "uglifyjs-webpack-plugin": "^1.2.5",
+    "url-loader": "^1.0.1",
+    "webpack": "^3.1.0",
+    "webpack-bundle-size-analyzer": "^2.5.0",
+    "webpack-cli": "^2.0.4",
+    "worker-loader": "^1.1.1"
   },
   "config": {
     "deps_check_dir": ".deps_check"
