@@ -1,8 +1,7 @@
 import React from "react"
 import PropTypes from "prop-types"
-import { Collapse } from "react-collapse"
 import { presets } from "react-motion"
-import ObjectInspector from "react-object-inspector"
+import ObjectInspector from "react-inspector"
 import Perf from "react-addons-perf"
 
 export default class Debug extends React.Component {
@@ -25,9 +24,11 @@ export default class Debug extends React.Component {
 
   render() {
 
-    let { getState } = this.props
+    let { getState, getComponent } = this.props
 
     window.props = this.props
+
+    const Collapse = getComponent("Collapse")
 
     return (
       <div className="info">
@@ -47,6 +48,7 @@ export default class Debug extends React.Component {
 }
 
 Debug.propTypes = {
-  getState: PropTypes.func.isRequired
+  getState: PropTypes.func.isRequired,
+  getComponent: PropTypes.func.isRequired,
 }
 
