@@ -36,11 +36,11 @@ export default class Parameters extends Component {
 
   onChange = ( param, value, isXml ) => {
     let {
-      specActions: { changeParam },
+      specActions: { changeParamByIdentity },
       onChangeKey,
     } = this.props
 
-    changeParam( onChangeKey, param.get("name"), param.get("in"), value, isXml)
+    changeParamByIdentity(onChangeKey, param, value, isXml)
   }
 
   onChangeConsumesWrapper = ( val ) => {
@@ -65,7 +65,8 @@ export default class Parameters extends Component {
       fn,
       getComponent,
       getConfigs,
-      specSelectors,
+      specSelectors, 
+      specActions,
       pathMethod
     } = this.props
 
@@ -101,11 +102,13 @@ export default class Parameters extends Component {
                       specPath={specPath.push(i.toString())}
                       getComponent={ getComponent }
                       getConfigs={ getConfigs }
-                      param={ parameter }
+                      rawParam={ parameter }
+                      param={ specSelectors.parameterWithMetaByIdentity(pathMethod, parameter) }
                       key={ `${parameter.get( "in" )}.${parameter.get("name")}` }
                       onChange={ this.onChange }
                       onChangeConsumes={this.onChangeConsumesWrapper}
                       specSelectors={ specSelectors }
+                      specActions={specActions}
                       pathMethod={ pathMethod }
                       isExecute={ isExecute }/>
                   )).toArray()
