@@ -21,8 +21,8 @@ export default class TryItOutButton extends React.Component {
     return (
       <div className="try-out">
         {
-          enabled ? <button className="btn try-out__btn cancel" onClick={ onTryoutClick }>Cancel</button>
-                  : <button className="btn try-out__btn" onClick={ onCancelClick }>Try it out </button>
+          enabled ? <button className="btn try-out__btn cancel" onClick={ onCancelClick }>Cancel</button>
+                  : <button className="btn try-out__btn" onClick={ onTryoutClick }>Try it out </button>
         }
       </div>
     )
