@@ -33,7 +33,8 @@ export default class HttpAuth extends React.Component {
     let { onChange } = this.props
     let { value, name } = e.target
 
-    let newValue = this.state.value || {}
+    let newValue = Object.assign({}, this.state.value)
+
     if(name) {
       newValue[name] = value
     } else {
@@ -53,7 +54,7 @@ export default class HttpAuth extends React.Component {
     const Markdown = getComponent( "Markdown" )
     const JumpToPath = getComponent("JumpToPath", true)
 
-    const scheme = schema.get("scheme")
+    const scheme = (schema.get("scheme") || "").toLowerCase()
     let value = this.getValue()
     let errors = errSelectors.allErrors().filter( err => err.get("authId") === name)
 
@@ -125,7 +126,7 @@ export default class HttpAuth extends React.Component {
     )
     }
   return <div>
-    <em><b>{name}</b> HTTP authentication: unsupported or missing scheme</em>
+    <em><b>{name}</b> HTTP authentication: unsupported scheme {`'${scheme}'`}</em>
   </div>
   }
 }
