@@ -1,4 +1,4 @@
-import { objectify, isFunc, normalizeArray } from "core/utils"
+import { objectify, isFunc, normalizeArray, deeplyStripKey } from "core/utils"
 import XML from "xml"
 import memoizee from "memoizee"
 
@@ -29,8 +29,14 @@ export const sampleFromSchema = (schema, config={}) => {
   let { type, example, properties, additionalProperties, items } = objectify(schema)
   let { includeReadOnly, includeWriteOnly } = config
 
-  if(example !== undefined)
-    return example
+
+  if(example !== undefined) {
+    return deeplyStripKey(example, "$$ref", (val) => {
+      // do a couple of quick sanity tests to ensure the value
+      // looks like a $$ref that swagger-client generates.
+      return typeof val === "string" && val.indexOf("#") > -1
+    })
+  }
 
   if(!type) {
     if(properties) {
@@ -46,10 +52,13 @@ export const sampleFromSchema = (schema, config={}) => {
     let props = objectify(properties)
     let obj = {}
     for (var name in props) {
-      if ( props[name].readOnly && !includeReadOnly ) {
+      if ( props[name] && props[name].deprecated ) {
         continue
       }
-      if ( props[name].writeOnly && !includeWriteOnly ) {
+      if ( props[name] && props[name].readOnly && !includeReadOnly ) {
+        continue
+      }
+      if ( props[name] && props[name].writeOnly && !includeWriteOnly ) {
         continue
       }
       obj[name] = sampleFromSchema(props[name], config)
@@ -69,6 +78,14 @@ export const sampleFromSchema = (schema, config={}) => {
   }
 
   if(type === "array") {
+    if(Array.isArray(items.anyOf)) {
+      return items.anyOf.map(i => sampleFromSchema(i, config))
+    }
+
+    if(Array.isArray(items.oneOf)) {
+      return items.oneOf.map(i => sampleFromSchema(i, config))
+    }
+
     return [ sampleFromSchema(items, config) ]
   }
 
@@ -202,7 +219,9 @@ export const sampleXmlFromSchema = (schema, config={}) => {
           || enumAttrVal || primitive(props[propName])
       } else {
         props[propName].xml.name = props[propName].xml.name || propName
-        props[propName].example = props[propName].example !== undefined ? props[propName].example : example[propName]
+        if(props[propName].example === undefined && example[propName] !== undefined) {
+          props[propName].example = example[propName]
+        }
         let t = sampleXmlFromSchema(props[propName])
         if (Array.isArray(t)) {
           res[displayName] = res[displayName].concat(t)
