@@ -2,7 +2,7 @@
 import React from "react"
 import expect, { createSpy } from "expect"
 import { shallow } from "enzyme"
-import { fromJS } from "immutable"
+import { fromJS, Map } from "immutable"
 import Models from "components/models"
 import ModelCollpase from "components/model-collapse"
 import ModelComponent from "components/model-wrapper"
@@ -19,12 +19,14 @@ describe("<Models/>", function(){
     },
     specSelectors: {
       isOAS3: () => false,
+      specJson: () => Map(),
       definitions: function() {
         return fromJS({
           def1: {},
           def2: {}
         })
-      }
+      },
+      specResolvedSubtree: () => {}
     },
     layoutSelectors: {
       isShown: createSpy()
