@@ -5,6 +5,7 @@ import System from "core/system"
 import { fromJS } from "immutable"
 import { render } from "enzyme"
 import ViewPlugin from "core/plugins/view/index.js"
+import filterPlugin from "core/plugins/filter/index.js"
 import { connect, Provider } from "react-redux"
 
 describe("bound system", function(){
@@ -264,6 +265,22 @@ describe("bound system", function(){
 
   })
 
+  describe("fn", function() {
+
+    it("should return helper functions", function () {
+      // Given
+      const system = new System({
+        plugins: [
+          filterPlugin
+        ]
+      })
+
+      // When
+      const fn = system.getSystem().fn.opsFilter
+      expect(typeof fn).toEqual("function")
+    })
+  })
+
   describe("selectors", function(){
 
     it("should have the first arg be the nested state, and all other args to follow", function(){
@@ -300,7 +317,7 @@ describe("bound system", function(){
 
     })
 
-    describe("when selector returns a funtcion", function(){
+    describe("when selector returns a function", function(){
 
       it("should pass the system to that function", function(){
 
@@ -677,6 +694,36 @@ describe("bound system", function(){
       })
 
       system.register([MyPlugin])
+
+      // When
+      var res = system.getSystem().wow()
+      expect(res).toEqual("so selective")
+    })
+  })
+
+  describe("rootInjects", function() {
+    it("should attach a rootInject function as an instance method", function() {
+      // This is the same thing as the `afterLoad` tests, but is here for posterity
+      
+      // Given
+      const system = new System({
+        plugins: [
+          {
+            afterLoad(system) {
+              this.rootInjects.wow = system.dogeSelectors.wow
+            },
+            statePlugins: {
+              doge: {
+                selectors: {
+                  wow: () => (system) => {
+                    return "so selective"
+                  }
+                }
+              }
+            }
+          }
+        ]
+      })
 
       // When
       var res = system.getSystem().wow()
